Add a service model
This breaks up the app into multiple units (again inspired by Android) that have
a well-defined (albeit restrictive) protocol for interactions. A service takes
over one particular aspect of the app (whatever that means) and has a lifecycle
under which that aspect of the app is managed.

A service can be loaded, started, stopped, and unloaded. It may be started again
once stopped, but it may not be loaded again once unloaded. This design decision
directly stems from the non-restartability of the CPython VM. So, the lifetime
of the CPython VM will be bounded by the load status of the "python" service,
and similar patterns will exist for other services.

Right now, empty skeletons for the "console", "face", "monitor", "python", and
"speech" exist. These will be filled in with the respective developments.

The "console" service will manage the terminal-based console user interface.

The "face" service will manage face recognition (moving out of Python).

The "monitor" service will manage the OpenGL visualization.

The "python" service will host the CPython virtual machine.

The "speech" service will manage speech recognition (also moving out of Python).

The face recognition and speech recognition were both implemented in Python,
originally, but the global interpreter lock has forced us to move compute-
intensive stuff out of Python and into native code. The robots are lagging
because of contention over the Python heap, as the face recognition script was
running in a background thread and using Python objects (on the heap). First and
foremost, this was only a problem because our computer's single-thread perf is
sooooo slow. BUT, it has four cores that are just begging for utilization, and
the GIL is not going to let us do that.

And no, I don't want multiple processes. We don't need multiple processes.